backend/requirements.txt
fastapi==0.119.0
uvicorn==0.37.0
sqlalchemy==2.0.44
psycopg2-binary==2.9.11
pymongo==4.15.3
python-multipart==0.0.20
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-dotenv==1.1.1
openai==2.4.0
groq==0.12.0
spacy==3.7.2
pydantic==2.12.2
email-validator==2.1.1



